project/requirements.txt
absl-py==0.9.0
appnope==0.1.0
argcomplete==1.12.1
astunparse==1.6.3
backcall==0.2.0
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cycler==0.10.0
decorator==4.4.2
gast==0.3.3
google-auth==1.20.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.30.0
h5py==2.10.0
idna==2.10
ipykernel==5.3.4
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.2
joblib==0.16.0
jupyter-client==6.1.6
jupyter-core==4.6.3
kiwisolver==1.2.0
Markdown==3.2.2
matplotlib==3.3.2
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.4
pandas==1.0.5
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
pipx==0.15.5.1
prompt-toolkit==3.0.5
protobuf==3.12.4
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.6.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
pyzmq==19.0.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==0.23.1
scipy==1.4.1
six==1.15.0
termcolor==1.1.0
threadpoolctl==2.1.0
tornado==6.0.4
traitlets==4.3.3
userpath==1.4.1
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
fastapi==0.60.1
plotly==4.9.0
uvicorn==0.11.8
python-dotenv==0.14.0
psycopg2-binary==2.8.6
aiofiles==0.6.0

